refact: simplify MilestoneCard and FooterDetails components by removing unnecessary props and console logs, enhancing code clarity and maintainability
